Disable automatic login when on AWS
AWS provides a locally accessible http api with information about
the instance. Check if this web server exists, and if so disable the
automatic login in gdm.

Add the AWS check as a pre-start step in the gdm service file. A one-
second delay will occur on all non-aws systems while the connection
times out.